refactor(pi-conductor): remove legacy worker state
