+++ packages/pi-conductor/__tests__/artifact-roots.test.ts
@@ -56,11 +56,8 @@ describe("conductor artifact root classification", () => {
           worktreePath: workerRoot,
           sessionFile: null,
           runtime: { backend: "session_manager" as const, sessionId: null, lastResumedAt: null },
-          currentTask: null,
           lifecycle: "idle" as const,
           recoverable: false,
-          lastRun: null,
-          summary: { text: null, updatedAt: null, stale: false },
           pr: { url: null, number: null, commitSucceeded: false, pushSucceeded: false, prCreationAttempted: false },
           createdAt: new Date().toISOString(),
           updatedAt: new Date().toISOString(),
