+++ packages/pi-conductor/__tests__/dispatch.test.ts
@@ -40,11 +40,8 @@ describe("conductor backend dispatch wrapper", () => {
           worktreePath: repoRoot,
           sessionFile: join(repoRoot, "session.jsonl"),
           runtime: { backend: "session_manager", sessionId: null, lastResumedAt: null },
-          currentTask: null,
           lifecycle: "idle",
           recoverable: false,
-          lastRun: null,
-          summary: { text: null, updatedAt: null, stale: false },
           pr: { url: null, number: null, commitSucceeded: false, pushSucceeded: false, prCreationAttempted: false },
           createdAt: now,
           updatedAt: now,
