+++ packages/pi-conductor/__tests__/index.test.ts
@@ -1,80 +1,75 @@
 import { readFileSync } from "node:fs";
 import { join } from "node:path";
-import { afterEach, describe, expect, it } from "vitest";
+import { describe, expect, it } from "vitest";
 import conductorExtension from "../extensions/index.js";
 
-function collectToolNames(): string[] {
-  const names: string[] = [];
+function collectRegistration(): { commands: string[]; tools: string[] } {
+  const commands: string[] = [];
+  const tools: string[] = [];
   const fakePi = {
-    registerCommand: () => undefined,
-    registerTool: (tool: { name: string }) => names.push(tool.name),
+    registerCommand: (name: string) => commands.push(name),
+    registerTool: (tool: { name: string }) => tools.push(tool.name),
   };
   conductorExtension(fakePi as never);
-  return names;
+  return { commands, tools };
 }
 
 describe("pi-conductor extension", () => {
-  afterEach(() => {
-    delete process.env.PI_CONDUCTOR_ENABLE_LEGACY_WORKER_TOOLS;
-  });
   it("registers the main conductor command group", () => {
-    const extension = readFileSync(join(__dirname, "../extensions/index.ts"), "utf-8");
-    expect(extension).toContain('registerCommand("conductor"');
+    expect(collectRegistration().commands).toEqual(["conductor"]);
   });
 
   it("registers resource-oriented conductor tools", () => {
-    const extension = readFileSync(join(__dirname, "../extensions/index.ts"), "utf-8");
-    expect(extension).toContain('name: "conductor_get_project"');
-    expect(extension).toContain('name: "conductor_list_objectives"');
-    expect(extension).toContain('name: "conductor_get_objective"');
-    expect(extension).toContain('name: "conductor_create_objective"');
-    expect(extension).toContain('name: "conductor_update_objective"');
-    expect(extension).toContain('name: "conductor_refresh_objective_status"');
-    expect(extension).toContain('name: "conductor_plan_objective"');
-    expect(extension).toContain('name: "conductor_link_task_to_objective"');
-    expect(extension).toContain('name: "conductor_list_workers"');
-    expect(extension).toContain('name: "conductor_list_tasks"');
-    expect(extension).toContain('name: "conductor_get_task"');
-    expect(extension).toContain('name: "conductor_list_runs"');
-    expect(extension).toContain('name: "conductor_list_gates"');
-    expect(extension).toContain('name: "conductor_create_worker"');
-    expect(extension).toContain('name: "conductor_create_task"');
-    expect(extension).toContain('name: "conductor_update_task"');
-    expect(extension).toContain('name: "conductor_assign_task"');
-    expect(extension).toContain('name: "conductor_delegate_task"');
-    expect(extension).not.toContain('name: "conductor_child_progress"');
-    expect(extension).not.toContain('name: "conductor_child_complete"');
-    expect(extension).not.toContain('name: "conductor_child_create_gate"');
-    expect(extension).toContain('name: "conductor_create_gate"');
-    expect(extension).toContain('name: "conductor_resolve_gate"');
-    expect(extension).not.toContain('name: "conductor_resolve_gate_as_human"');
-    expect(extension).toContain('name: "conductor_start_task_run"');
-    expect(extension).toContain('name: "conductor_run_task"');
-    expect(extension).toContain('name: "conductor_cancel_task_run"');
-    expect(extension).toContain('name: "conductor_retry_task"');
-    expect(extension).toContain('name: "conductor_list_events"');
-    expect(extension).toContain('name: "conductor_list_artifacts"');
-    expect(extension).toContain('name: "conductor_backend_status"');
-    expect(extension).toContain('name: "conductor_reconcile_project"');
-    expect(extension).toContain('name: "conductor_next_actions"');
-    expect(extension).toContain('name: "conductor_project_brief"');
-    expect(extension).toContain('name: "conductor_task_brief"');
-    expect(extension).toContain('name: "conductor_resource_timeline"');
-    expect(extension).toContain('name: "conductor_run_next_action"');
-    expect(extension).toContain('name: "conductor_scheduler_tick"');
-    expect(extension).toContain('name: "conductor_schedule_objective"');
-    expect(extension).toContain('name: "conductor_assess_task"');
-    expect(extension).toContain('name: "conductor_read_artifact"');
-    expect(extension).toContain('name: "conductor_objective_dag"');
-    expect(extension).toContain('name: "conductor_prepare_human_review"');
-    expect(extension).toContain('name: "conductor_diagnose_blockers"');
-    expect(extension).toContain('name: "conductor_build_evidence_bundle"');
-    expect(extension).toContain('name: "conductor_check_readiness"');
-    expect(extension).toContain('name: "conductor_recover_worker"');
-    expect(extension).toContain('name: "conductor_cleanup_worker"');
-    expect(extension).toContain('name: "conductor_commit_worker"');
-    expect(extension).toContain('name: "conductor_push_worker"');
-    expect(extension).toContain('name: "conductor_create_worker_pr"');
+    const expectedTools = [
+      "conductor_get_project",
+      "conductor_list_objectives",
+      "conductor_get_objective",
+      "conductor_create_objective",
+      "conductor_update_objective",
+      "conductor_refresh_objective_status",
+      "conductor_plan_objective",
+      "conductor_link_task_to_objective",
+      "conductor_reconcile_project",
+      "conductor_run_next_action",
+      "conductor_schedule_objective",
+      "conductor_scheduler_tick",
+      "conductor_assess_task",
+      "conductor_read_artifact",
+      "conductor_objective_dag",
+      "conductor_prepare_human_review",
+      "conductor_diagnose_blockers",
+      "conductor_resource_timeline",
+      "conductor_task_brief",
+      "conductor_project_brief",
+      "conductor_next_actions",
+      "conductor_build_evidence_bundle",
+      "conductor_check_readiness",
+      "conductor_backend_status",
+      "conductor_list_events",
+      "conductor_list_artifacts",
+      "conductor_list_workers",
+      "conductor_list_tasks",
+      "conductor_get_task",
+      "conductor_list_runs",
+      "conductor_list_gates",
+      "conductor_create_worker",
+      "conductor_create_task",
+      "conductor_update_task",
+      "conductor_assign_task",
+      "conductor_delegate_task",
+      "conductor_create_gate",
+      "conductor_resolve_gate",
+      "conductor_start_task_run",
+      "conductor_run_task",
+      "conductor_cancel_task_run",
+      "conductor_retry_task",
+      "conductor_recover_worker",
+      "conductor_cleanup_worker",
+      "conductor_commit_worker",
+      "conductor_push_worker",
+      "conductor_create_worker_pr",
+    ];
+    expect([...collectRegistration().tools].sort()).toEqual([...expectedTools].sort());
   });
 
   it("packages conductor workflow skills", () => {
@@ -86,21 +81,5 @@ describe("pi-conductor extension", () => {
     expect(packageJson.pi.skills).toEqual(["./skills"]);
     expect(gateReviewSkill).toContain("/conductor human dashboard");
     expect(orchestrationSkill).toContain('policy: "execute"');
-  });
-
-  it("hides legacy worker tools unless compatibility is enabled", () => {
-    const names = collectToolNames();
-    expect(names).toContain("conductor_get_project");
-    expect(names).not.toContain("conductor_start");
-    expect(names).not.toContain("conductor_run");
-    expect(names).not.toContain("conductor_pr_create");
-  });
-
-  it("registers legacy worker tools when compatibility is enabled", () => {
-    process.env.PI_CONDUCTOR_ENABLE_LEGACY_WORKER_TOOLS = "1";
-    const names = collectToolNames();
-    expect(names).toContain("conductor_start");
-    expect(names).toContain("conductor_run");
-    expect(names).toContain("conductor_pr_create");
   });
 });
