+++ packages/pi-conductor/__tests__/llm-controls.test.ts
@@ -121,11 +121,8 @@ describe("LLM autonomous control helpers", () => {
       worktreePath: "/tmp/worktree",
       sessionFile: "/tmp/session",
       runtime: { backend: "session_manager" as const, sessionId: null, lastResumedAt: null },
-      currentTask: null,
       lifecycle: "idle" as const,
       recoverable: false,
-      lastRun: null,
-      summary: { text: null, updatedAt: null, stale: false },
       pr: { url: null, number: null, commitSucceeded: false, pushSucceeded: false, prCreationAttempted: false },
       createdAt: new Date().toISOString(),
       updatedAt: new Date().toISOString(),
