+++ packages/pi-conductor/__tests__/llm-hardening.test.ts
@@ -40,11 +40,8 @@ describe("LLM hardening follow-ups", () => {
       worktreePath: repoRoot,
       sessionFile: join(repoRoot, "session.jsonl"),
       runtime: { backend: "session_manager" as const, sessionId: null, lastResumedAt: null },
-      currentTask: null,
       lifecycle: "idle" as const,
       recoverable: false,
-      lastRun: null,
-      summary: { text: null, updatedAt: null, stale: false },
       pr: { url: null, number: null, commitSucceeded: false, pushSucceeded: false, prCreationAttempted: false },
       createdAt: new Date().toISOString(),
       updatedAt: new Date().toISOString(),
