+++ packages/pi-conductor/__tests__/objective-scheduler.test.ts
@@ -41,11 +41,8 @@ describe("objective DAG scheduler", () => {
         worktreePath: repoRoot,
         sessionFile: join(repoRoot, `session-${index + 1}.jsonl`),
         runtime: { backend: "session_manager" as const, sessionId: null, lastResumedAt: null },
-        currentTask: null,
         lifecycle: "idle" as const,
         recoverable: false,
-        lastRun: null,
-        summary: { text: null, updatedAt: null, stale: false },
         pr: { url: null, number: null, commitSucceeded: false, pushSucceeded: false, prCreationAttempted: false },
         createdAt: now,
         updatedAt: now,
