+++ packages/pi-conductor/__tests__/sessions.test.ts
@@ -1,8 +1,8 @@
 import { execSync } from "node:child_process";
-import { existsSync, readFileSync, rmSync, writeFileSync } from "node:fs";
+import { existsSync, rmSync, writeFileSync } from "node:fs";
 import { join } from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
-import { createWorkerSessionRuntime, resumeWorkerSessionRuntime } from "../extensions/runtime.js";
+import { createWorkerSessionRuntime } from "../extensions/runtime.js";
 
 function requireValue<T>(value: T | null | undefined, message: string): T {
   if (value == null) {
@@ -34,16 +34,5 @@ describe("session linkage", () => {
     const sessionFile = requireValue((await createWorkerSessionRuntime(repoDir)).sessionFile, "session file missing");
     expect(sessionFile).toBeTruthy();
     expect(existsSync(sessionFile)).toBe(true);
-  });
-
-  it("opens an existing worker session through the runtime layer and records a resume entry", async () => {
-    const sessionFile = requireValue((await createWorkerSessionRuntime(repoDir)).sessionFile, "session file missing");
-    const runtime = await resumeWorkerSessionRuntime(sessionFile);
-
-    expect(runtime.sessionFile).toBe(sessionFile);
-    expect(runtime.sessionId).toBeTruthy();
-    expect(runtime.backend).toBe("session_manager");
-    expect(runtime.lastResumedAt).toBeTruthy();
-    expect(readFileSync(sessionFile, "utf-8")).toContain("pi-conductor_runtime_resume");
   });
 });
